Apply the patch from Dariusz Kordonski (dariusz.kordonski@googlemail.com) to the 1.2.2 branch. This gets the idea-plugin working on IDEA version 9.
Fixed the ant build as well. This may be broken on Misko's machine since IDEA9 isn't present, and the hudson config doesn't include a path to the idea.jar for use in Ant. Someone will need to fix that :)

Right now, I can run the plugin in the sandbox IDE and successfully start the server, capture a browser, and run tests. So we are basically working. Next step is to fix a GUI call off the AWT thread, do some more testing, and cut a release of the plugin.

Update issue 82