Improve the IntelliJ project setup, so that the core library code is depended on as source, not as a jar.
The previous setup made it hard to navigate from plugin code back to core library sources, and making changes in the core library required a separate project setup, an ant build, and project refresh to pick up changes. This was required because the TestRunner's classpath needs to list all the library dependencies of JsTD, and we brought those in by using the jarjar'ed jstd.jar which includes them.
 
 By constructing that classpath more carefully, we now find the jars in the plugin's directory, and also the compiled classes from the IDE, and it's now possible to have the plugin and the core library as two modules in the same project.